Couple of small fixes
Fixed some issues with the animator state machine and the animations in general that lead to weird behavior in certain cases. The opponent now gets stunned for a short duration after being hit.

Simplified hitbox of the projectile and Bow ammo prefabs.

Rotation of the opponents is now no longer controlled by the NavMeshAgent during combat. This will enable the opponents to move sideways and backwards while still looking at the player later on.