Further improvements to enemy animations and Ai
Enemies now use a blend tree for movement animations in all four directions, meaning they can all be bundled in one state and interpolated between simultaenously. Enemies now also need a clear line of sight towards the player to be able to see him, which currently means the player has to be within 20 units of the enemy, inside of a frontal cone of 120 degrees, and there can be no wall blocking the line of sight. Enemies will also become aware of the player when he is within 5 units of them, unless the player is sneaking. When the enemies lose sight of the player, they will continue towards his last known position, remain there for a short while, and then return to their nav points inside of their spawn room.